Moved API data fetching to the App component
