+++ src/App.js
@@ -1,13 +1,36 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { Switch, Route } from "react-router-dom";
+import { connect } from "react-redux";
 
 import Header from "./components/Header";
 import Restaurant from "./components/Restaurant";
 import Restaurants from "./components/Restaurants";
+import { actionTypes } from "./redux-store/actionTypes";
 
 import "./index.scss";
 
-const App = () => {
+const apiLink = "https://redi-final-restaurants.herokuapp.com/restaurants";
+
+const App = ({ dispatch }) => {
+  useEffect(() => {
+    async function fetchData() {
+      dispatch({ type: actionTypes.startLoading, payload: { error: false } });
+      try {
+        const response = await fetch(apiLink);
+        const data = await response.json();
+        const restaurants = data.results;
+        // Dispatching the list of all restaurants to the redux store
+        dispatch({
+          type: actionTypes.loadSuccess,
+          payload: { restaurants },
+        });
+      } catch (error) {
+        dispatch({ type: actionTypes.loadError, payload: { error } });
+      }
+    }
+
+    fetchData();
+  }, [dispatch]);
   return (
     <div className="wrapper">
       <Header />
@@ -25,4 +48,8 @@ const App = () => {
   );
 };
 
-export default App;
+function mapReduxStateToProps(reduxState) {
+  return { restaurants: reduxState.restaurants };
+}
+
+export default connect(mapReduxStateToProps)(App);
